New plugin to return random video from a youtube playlist
It's called GTAPlugin, but it's actually quite generic.  It doesn't use
the youtube plugin RSS feed because I wanted to use this on an unlisted
playlist, and the RSS feed for unlisted playlists don't return any
videos.

So I'm using an regular expression.  You can judge me for this, but not
too badly.